# portage: fix changed_use and newuse not triggering rebuilds (#6008)