--- a/changelogs/fragments/6548-portage-changed_use-newuse.yml
+++ b/changelogs/fragments/6548-portage-changed_use-newuse.yml
@@ -1,3 +1,3 @@
 ---
 bugfixes:
-  - portage - fix ``changed_use`` and ``newuse`` not triggering rebuilds (https://github.com/ansible-collections/community.general/issues/6008)
+  - portage - fix ``changed_use`` and ``newuse`` not triggering rebuilds (https://github.com/ansible-collections/community.general/issues/6008, https://github.com/ansible-collections/community.general/pull/6548).